adds recent project based on fcc_username
