+++ client/src/Member/MemberCard.js
@@ -4,11 +4,6 @@ import './Member.css';
 import MemberStack from './MemberStack';
 import MemberMeetup from './MemberMeetup';
 
-const Xray = require('x-ray');
-const x = Xray();
-let projects = [];
-const re = /^(https:\/\/freecodecamp.org)|(https:\/\/freecodecamp.com)/
-
 const MemberCard = ({ props }) => {
   const name = props.name || "na"
   const surname = props.surname || "na"
@@ -18,27 +13,13 @@ const MemberCard = ({ props }) => {
   const bio = props.bio || "na"
   const stack = props.stack || []
   
-  let recent = "Loading..."
+  const recent_url = props.fcc.fcc_recent || "#"
+  const recent_slug = recent_url.match(/[^/]+(?=\/$|$)/).toString().replace(/-/g, ' ');
   const fcc_forum_stats = "Loading..."
   const github = props.contact.github ? props.contact.github : "mrslwiseman";
   const twitter = props.contact.twitter ? props.contact.twitter : "mrslwiseman";
   const fcc_username = props.fcc.fcc_username ? props.fcc.fcc_username : "mrslwiseman";
   
-  const scrape = (fcc_username) => {
-    return new Promise((yeah, nah) => {
-      x(`https://freecodecamp.org/${fcc_username}`, ['a@href'])((err, data) => {
-        if (err) console.log(err);
-        let links = data.toString().split(',');
-        for (let link of links) {
-          if (!link.match(re)) {
-            projects.push(link);
-          }
-        }
-        (projects.length > 1) ? yeah(projects[projects.length-1]) : nah('Projects not found');
-      });
-
-    })
-
   return (
     <div className="col-sm-4 col-lg-3">
       <div className="panel panel-danger">
@@ -61,7 +42,7 @@ const MemberCard = ({ props }) => {
         <div>
           <h4>Free Code Camp Details:</h4>
           <h5>Recent Project:</h5> 
-          {this.scrape(fcc_username)}
+          {recent_url != '#' ? <a href={`${recent_url}`}>{`${recent_slug}`}</a> : "No project found!"}
           <h5>FCC Forum Stats:</h5> 
           {fcc_forum_stats}
 
@@ -88,7 +69,6 @@ const MemberCard = ({ props }) => {
       </div>
     </div>
   )
-}
 }
